requirements.txt
dj-database-url==0.5.0
Django==1.11.1
django-bootstrap-form==3.2.1
django-bootstrap4==0.0.7
django-crispy-forms==1.6.0
django-heroku==0.3.1
django-import-export==1.0.1
gunicorn==19.6.0
Pillow==3.1.0
psycopg2==2.7.4
pytz==2017.2
whitenoise==3.2
django-tinymce
